refactor ui panels
