+++ openfoam_cbp/server/index.js
@@ -14,12 +14,11 @@ const hostname = '0.0.0.0'
 const localhost = 'localhost'
 const DIST_DIR = path.join(__dirname, '../dist');
 const HTML_FILE = path.join(__dirname, '../dist/index.html');
-
 app.use(express.static(DIST_DIR));
 app.use(express.urlencoded({extended: true}));
 app.use(express.json({ limit: '50mb' }));
 app.use(session({
-  secret: "8e3ce5c7f1b15ab25f5690413740db0df59f693c",
+  secret: require('crypto').randomBytes(48).toString('hex'),
   resave: false,
   saveUninitialized: true
 }))
