+++ openfoam_cbp/src/components/panels/editpanel/uipanels/fvschemes.jsx
@@ -1,0 +1,1 @@
+export const FvSchemes = ({ast, editor}) => {return null}
