+++ openfoam_cbp/src/components/panels/editpanel/uipanels/fvsolution.jsx
@@ -1,0 +1,3 @@
+export const FvSolution = ({ast, editor}) => {
+    return null
+}
